+++ src/main.tsx
@@ -3,8 +3,6 @@ import { createRoot } from "react-dom/client";
 import { ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
 import { theme } from "@shared/theme/theme";
-import "./index.css";
-import "./assets/styles/global.css";
 import App from "./App.tsx";
 
 createRoot(document.getElementById("root")!).render(